refactor endpoint service code
