+++ packages/browser/src/Component/Widget/Server/SPARQL/Service/SparqlEndpointService.js
